Implement registration wizard in profile_edit view.
When user.is_matrimony_registration_complete is False, use the profile_edit forms in wizard mode.

In the wizard mode, you will not be able to traverse to next tabs without filling the current tab.
You will see 'Next' in place of 'Save' in the submit buttons, and on proper form submit, you
will be taken to the next section.

On submitting the last of forms, user.is_matrimony_registration_complete will be set to True,
and you will be able to freely edit any of the profile forms.